+++ src/main/java/com/epita/mti/datemine/data/Entity/Project.java
@@ -28,7 +28,7 @@ public class Project extends AbstractEntity {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column
-    @Getter @Setter private Integer id;
+    @Getter @Setter private Long id;
 
     /**
      * The project name.
